Added ctf_perimetersphere entity
Nearly identical entity to ctf_constructsphere, which is invisible and scales with the ctf_buildzonescale console command. Local prints called from inside the entity work in-game, so testing will be required for ensuring that the prints do not call outside of the invisible barrier, and that it is not removed when the match begins.