Added functional timer and perimeter sphere call
Added and tested a functional timer, which awards money to the player every five seconds. This code is confirmed to have worked. As such, I have moved it into the (MatchHasBegun) condition, where the menu restriction takes place when the round begins, since that is when we want the timer to start.

I have also added initialization for the ctf_perimeter_sphere entity which is confirmed to be working in game, see the details on that push for more information in the entities folder.