Added dynamic cost variables
Added cVars, implemented dynamic counts for kills and the timer that awards the player money passively. Removed non-functional giveMoney command (Add in later) and added an NPC kill function that is identical to the player kill function for singleplayer testing.

All have been reported as bug free at this time.