Improved simfphys implementation
Added cost prices to the items in the meny, added in a partially working simfphys population loop for the vehicles tab (See line 192-203). Added the spitfire for purchase from the menu. Cost concatenation is not complete for the vehicles tab and is thus simplified to only the entity name. Fixed model icons for other entities.